+++ Realization/frontend/czechidm-acc/src/services/SystemService.js
@@ -14,8 +14,15 @@ class SystemService extends Services.AbstractService {
     return system.name;
   }
 
-  // dto
   supportsPatch() {
+    return false;
+  }
+
+  getGroupPermission() {
+    return 'SYSTEM';
+  }
+
+  supportsAuthorization() {
     return false;
   }